// OneDrive/Desktop/fuck_elections/typechain-types/VotoMVP.ts
export interface VotoMVPInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "adicionarCandidato"
      | "candidatos"
      | "getTotalCandidatos"
      | "jaVotou"
      | "owner"
      | "totalVotes"
      | "votar"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "CandidatoAdicionado" | "VotoRegistrado"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "adicionarCandidato",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "candidatos",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getTotalCandidatos",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "jaVotou",
    values: [AddressLike]
  ): string;
  encodeFunctionData(functionFragment: "owner", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "totalVotes",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "votar", values: [BigNumberish]): string;

  decodeFunctionResult(
    functionFragment: "adicionarCandidato",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "candidatos", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getTotalCandidatos",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "jaVotou", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "owner", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "totalVotes", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "votar", data: BytesLike): Result;
}

export namespace CandidatoAdicionadoEvent {
  export type InputTuple = [candidatoId: BigNumberish, nome: string];
  export type OutputTuple = [candidatoId: bigint, nome: string];
  export interface OutputObject {
    candidatoId: bigint;
    nome: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace VotoRegistradoEvent {
  export type InputTuple = [eleitor: AddressLike, candidatoId: BigNumberish];
  export type OutputTuple = [eleitor: string, candidatoId: bigint];
  export interface OutputObject {
    eleitor: string;
    candidatoId: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface VotoMVP extends BaseContract {
  connect(runner?: ContractRunner | null): VotoMVP;
  waitForDeployment(): Promise<this>;

  interface: VotoMVPInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  adicionarCandidato: TypedContractMethod<
    [_nome: string],
    [void],
    "nonpayable"
  >;

  candidatos: TypedContractMethod<
    [arg0: BigNumberish],
    [[string, bigint] & { nome: string; voteCount: bigint }],
    "view"
  >;

  getTotalCandidatos: TypedContractMethod<[], [bigint], "view">;

  jaVotou: TypedContractMethod<[arg0: AddressLike], [boolean], "view">;

  owner: TypedContractMethod<[], [string], "view">;

  totalVotes: TypedContractMethod<[], [bigint], "view">;

  votar: TypedContractMethod<
    [_candidatoId: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "adicionarCandidato"
  ): TypedContractMethod<[_nome: string], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "candidatos"
  ): TypedContractMethod<
    [arg0: BigNumberish],
    [[string, bigint] & { nome: string; voteCount: bigint }],
    "view"
  >;
  getFunction(
    nameOrSignature: "getTotalCandidatos"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "jaVotou"
  ): TypedContractMethod<[arg0: AddressLike], [boolean], "view">;
  getFunction(
    nameOrSignature: "owner"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "totalVotes"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "votar"
  ): TypedContractMethod<[_candidatoId: BigNumberish], [void], "nonpayable">;

  getEvent(
    key: "CandidatoAdicionado"
  ): TypedContractEvent<
    CandidatoAdicionadoEvent.InputTuple,
    CandidatoAdicionadoEvent.OutputTuple,
    CandidatoAdicionadoEvent.OutputObject
  >;
  getEvent(
    key: "VotoRegistrado"
  ): TypedContractEvent<
    VotoRegistradoEvent.InputTuple,
    VotoRegistradoEvent.OutputTuple,
    VotoRegistradoEvent.OutputObject
  >;

  filters: {
    "CandidatoAdicionado(uint256,string)": TypedContractEvent<
      CandidatoAdicionadoEvent.InputTuple,
      CandidatoAdicionadoEvent.OutputTuple,
      CandidatoAdicionadoEvent.OutputObject
    >;
    CandidatoAdicionado: TypedContractEvent<
      CandidatoAdicionadoEvent.InputTuple,
      CandidatoAdicionadoEvent.OutputTuple,
      CandidatoAdicionadoEvent.OutputObject
    >;

    "VotoRegistrado(address,uint256)": TypedContractEvent<
      VotoRegistradoEvent.InputTuple,
      VotoRegistradoEvent.OutputTuple,
      VotoRegistradoEvent.OutputObject
    >;
    VotoRegistrado: TypedContractEvent<
      VotoRegistradoEvent.InputTuple,
      VotoRegistradoEvent.OutputTuple,
      VotoRegistradoEvent.OutputObject
    >;
  };
}